Comment out the kv_score up-cast again

+++ python/sglang/srt/mem_cache/deepseek_v4_compress_state.py
@@ -56,7 +56,8 @@ class KVAndScore:
         # EAGLE/MTP spec-decode produces kv_score in bf16 (compressor.py) while
         # this buffer is float32 (SGLANG_DSV4_COMPRESS_STATE_DTYPE default). Up-cast
         # to the destination dtype so the index-put doesn't crash (Float dst / BF16 src).
-        self.kv_score[index] = value.kv_score.to(self.kv_score.dtype)
+        # self.kv_score[index] = value.kv_score.to(self.kv_score.dtype)
+        self.kv_score[index] = value.kv_score
 
     def clear(self):
         self.kv.zero_()
